Drop kestrel fully
