+++ OmniSharp.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26730.0
+VisualStudioVersion = 15.0.26730.3
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{2C348365-A9D8-459E-9276-56FC46AAEE31}"
 EndProject
@@ -48,8 +48,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OmniSharp.Roslyn.CSharp.Tes
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OmniSharp.Host", "src\OmniSharp.Host\OmniSharp.Host.csproj", "{A1B54CA5-7F82-4487-8150-B4CB3952A2AD}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OmniSharp", "src\OmniSharp\OmniSharp.csproj", "{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OmniSharp.DotNet", "src\OmniSharp.DotNet\OmniSharp.DotNet.csproj", "{07CDB87A-6906-4B04-9622-164FE04A3F9A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestUtility", "tests\TestUtility\TestUtility.csproj", "{9459D3EB-9D7F-437E-AC7A-245698B4D62F}"
@@ -59,10 +57,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OmniSharp.DotNetTest.Tests", "tests\OmniSharp.DotNetTest.Tests\OmniSharp.DotNetTest.Tests.csproj", "{F6195E7B-8002-42AE-BF86-E78DBC769A47}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OmniSharp.DotNet.Tests", "tests\OmniSharp.DotNet.Tests\OmniSharp.DotNet.Tests.csproj", "{A645D475-3448-4473-88CA-E3C3B31E33CA}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OmniSharp.Http", "src\OmniSharp.Http\OmniSharp.Http.csproj", "{296B8800-007B-41D5-9001-50A371B8B5F6}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OmniSharp.Http.Tests", "tests\OmniSharp.Http.Tests\OmniSharp.Http.Tests.csproj", "{DA207736-4F3A-4330-8048-2FCB170559F7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -230,18 +224,6 @@ Global
 		{A1B54CA5-7F82-4487-8150-B4CB3952A2AD}.Release|x64.Build.0 = Release|Any CPU
 		{A1B54CA5-7F82-4487-8150-B4CB3952A2AD}.Release|x86.ActiveCfg = Release|Any CPU
 		{A1B54CA5-7F82-4487-8150-B4CB3952A2AD}.Release|x86.Build.0 = Release|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Debug|x64.Build.0 = Debug|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Debug|x86.Build.0 = Debug|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Release|x64.ActiveCfg = Release|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Release|x64.Build.0 = Release|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Release|x86.ActiveCfg = Release|Any CPU
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94}.Release|x86.Build.0 = Release|Any CPU
 		{07CDB87A-6906-4B04-9622-164FE04A3F9A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{07CDB87A-6906-4B04-9622-164FE04A3F9A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{07CDB87A-6906-4B04-9622-164FE04A3F9A}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -302,30 +284,6 @@ Global
 		{A645D475-3448-4473-88CA-E3C3B31E33CA}.Release|x64.Build.0 = Release|Any CPU
 		{A645D475-3448-4473-88CA-E3C3B31E33CA}.Release|x86.ActiveCfg = Release|Any CPU
 		{A645D475-3448-4473-88CA-E3C3B31E33CA}.Release|x86.Build.0 = Release|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Debug|x64.Build.0 = Debug|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Debug|x86.Build.0 = Debug|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Release|x64.ActiveCfg = Release|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Release|x64.Build.0 = Release|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Release|x86.ActiveCfg = Release|Any CPU
-		{296B8800-007B-41D5-9001-50A371B8B5F6}.Release|x86.Build.0 = Release|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Debug|x64.Build.0 = Debug|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Debug|x86.Build.0 = Debug|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Release|x64.ActiveCfg = Release|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Release|x64.Build.0 = Release|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Release|x86.ActiveCfg = Release|Any CPU
-		{DA207736-4F3A-4330-8048-2FCB170559F7}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -344,14 +302,11 @@ Global
 		{194A88AE-1429-416B-86C1-BF6F7FF47404} = {35E025BF-BBB2-4FAC-9F4B-37CBA083EE47}
 		{2F4B72DA-FC27-4676-8E89-7BAC2EFDF750} = {35E025BF-BBB2-4FAC-9F4B-37CBA083EE47}
 		{A1B54CA5-7F82-4487-8150-B4CB3952A2AD} = {2C348365-A9D8-459E-9276-56FC46AAEE31}
-		{CD970E2E-CEC1-454D-AC7E-23D0000DCC94} = {2C348365-A9D8-459E-9276-56FC46AAEE31}
 		{07CDB87A-6906-4B04-9622-164FE04A3F9A} = {2C348365-A9D8-459E-9276-56FC46AAEE31}
 		{9459D3EB-9D7F-437E-AC7A-245698B4D62F} = {35E025BF-BBB2-4FAC-9F4B-37CBA083EE47}
 		{33A68F57-B234-4481-AEDE-03D51756B35A} = {2C348365-A9D8-459E-9276-56FC46AAEE31}
 		{F6195E7B-8002-42AE-BF86-E78DBC769A47} = {35E025BF-BBB2-4FAC-9F4B-37CBA083EE47}
 		{A645D475-3448-4473-88CA-E3C3B31E33CA} = {35E025BF-BBB2-4FAC-9F4B-37CBA083EE47}
-		{296B8800-007B-41D5-9001-50A371B8B5F6} = {2C348365-A9D8-459E-9276-56FC46AAEE31}
-		{DA207736-4F3A-4330-8048-2FCB170559F7} = {35E025BF-BBB2-4FAC-9F4B-37CBA083EE47}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4DD725CE-B49A-4151-8B77-BB33FE88E46E}
