--- src/OmniSharp.Http/OmniSharp.Http.csproj
@@ -1,24 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net46</TargetFramework>
-        <WarningsAsErrors>true</WarningsAsErrors>
-        <PreserveCompilationContext>true</PreserveCompilationContext>
-        <OutputType>Exe</OutputType>
-        <RuntimeIdentifiers>win7-x64;win7-x86</RuntimeIdentifiers>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.Diagnostics" Version="1.1.0" />
-        <PackageReference Include="Microsoft.AspNetCore.Server.Kestrel" Version="1.1.0" />
-        <PackageReference Include="Microsoft.AspNetCore.Hosting" Version="1.1.0" />
-        <PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="1.1.0" />
-        <ProjectReference Include="..\OmniSharp.Host\OmniSharp.Host.csproj" />
-        <ProjectReference Include="..\OmniSharp.Roslyn.CSharp\OmniSharp.Roslyn.CSharp.csproj" />
-        <ProjectReference Include="..\OmniSharp.DotNet\OmniSharp.DotNet.csproj" />
-        <ProjectReference Include="..\OmniSharp.DotNetTest\OmniSharp.DotNetTest.csproj" />
-        <ProjectReference Include="..\OmniSharp.MSBuild\OmniSharp.MSBuild.csproj" />
-        <ProjectReference Include="..\OmniSharp.Script\OmniSharp.Script.csproj" />
-    </ItemGroup>
-
-</Project>
