--- src/OmniSharp/OmniSharp.csproj
@@ -1,22 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net46</TargetFramework>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <PreserveCompilationContext>true</PreserveCompilationContext>
-    <OutputType>Exe</OutputType>
-    <RuntimeIdentifiers>win7-x64;win7-x86</RuntimeIdentifiers>
-  </PropertyGroup>
-
-    <ItemGroup>
-        <Content Include="config.json">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <None Include="app.config" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\OmniSharp.Stdio\OmniSharp.Stdio.csproj" />
-    </ItemGroup>
-
-</Project>
