--- src/OmniSharp/Program.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Threading;
-
-public class Program
-{
-    public static int Main(string[] args)
-    {
-        return OmniSharp.Stdio.Program.Main(args);
-    }
-}
