--- src/OmniSharp/config.json
@@ -1,17 +1,0 @@
-﻿{
-    "dotnet": {
-        "alias": "default",
-        "projects": "**/project.json",
-        "enablePackageRestore": true,
-        "packageRestoreTimeout": "180"
-    },
-    "formattingOptions": {
-        "newLine": "\n",
-        "useTabs": false,
-        "tabSize": 4,
-        "indentationSize": 4
-    },
-    "script": {
-        "enableScriptNuGetReferences" : false
-    }
-}
