--- tests/OmniSharp.Http.Tests/OmniSharp.Http.Tests.csproj
@@ -1,29 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net46</TargetFramework>
-    <WarningsAsErrors>true</WarningsAsErrors>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include="..\app.config" Link="app.config" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\OmniSharp.Http\OmniSharp.Http.csproj" />
-    <ProjectReference Include="..\TestUtility\TestUtility.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.0.0" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.2.0" />
-    <PackageReference Include="xunit" Version="2.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
-  </ItemGroup>
-
-</Project>
